Probably made LUT now more thread-safe. Modified linear_to_srgb to also be thread-safe. Added some changes for scalar case to respect new LUT datatype being uint32 (not uint8 anymore only to make linear_to_srgb thread-safe), appending static casts to uint8 where needed.

+++ include/alpha_hist/color_lut.hpp
@@ -3,6 +3,7 @@
 #include <cstdint>
 #include <vector>
 
+
 namespace alpha_hist {
 
     struct ColorLUT {
@@ -10,7 +11,7 @@ namespace alpha_hist {
 
         std::array<float, kSrgbSize> srgb_to_linear{};
 
-        std::vector<std::uint8_t> linear_to_srgb;
+        std::vector<std::uint32_t> linear_to_srgb;
         int linear_size = 0;
     };
